ITSE-44965: Potential fix for arbitrary code execution vulnerability
ITSE-44965 describes an arbitrary code execution vulnerability in squatch-js related to the auto popup widget functionality. By using the domain as specified in the _saasquatchExtra URL parameter to fetch widget HTML the code allows an attacker to insert arbitrary HTML (including Javascript) from an attacker domain directly into the DOM.

This fix removes the dependency on domain and tenantAlias as specified in _saasquatchExtra and instead uses the values from the global squatch-js config to navigate the _saasquatchExtra parameter to determine if a widget should be automaticallly loaded.